requirements.txt
pytest==5.4.3
requests==2.24.0
PyHamcrest==2.0.2
allure-pytest==2.8.17